(E1.0) Excel at Nodejs - Export into Static HTML App
had to just clone their repo in the end and copy over the contents of
the E1 desired state of the app to be able to get it to work, because
the app state at the end of Basics was not in line with what they
intended.

anyways, so far of E1 / 9, they've had me create the next.config.js
file, add the `export` script in package.json, isntall `serve` (npm),
and then export the app into static html and serve it with serve,
configured from the next.config.